LayoutTests/media/modern-media-controls/media-controller/media-controller-constructor-expected.txt

Testing the MediaController constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS createControls is defined.
PASS createControls is an instance of Function
PASS mediaController is an instance of MediaController
PASS mediaController.shadowRoot is shadowRoot
PASS mediaController.media is media
PASS mediaController.controls is an instance of MediaControls

PASS successfullyParsed is true

TEST COMPLETE
